requirements.txt
Django==1.7
djangorestframework==2.4.3
wsgiref==0.1.2
django-extensions==1.4.0
